refactor(plugins): temporaryName -> blockchainUtil
