+++ lib/blockchain_util.js
