+++ lib/plugins.js
@@ -17,7 +17,7 @@ var traderPlugin = null;
 var walletPlugin = null;
 var idVerifierPlugin = null;
 
-var temporaryName = null;
+var blockchainUtil = null;
 
 var currentlyUsedPlugins = {};
 
@@ -169,8 +169,8 @@ exports.configure = function configure(config) {
   );
 
   // NOTE: temp solution
-  if (temporaryName === null)
-    temporaryName = require('./temporary_name');
+  if (blockchainUtil === null)
+    blockchainUtil = require('./blockchain_util');
 };
 exports.getCachedConfig = function getCachedConfig() {
   return cachedConfig;
@@ -305,7 +305,7 @@ function _monitorAddress(address, cb) {
   var interval = 300;   // TODO make config
 
   function checkAddress(_cb) {
-    temporaryName.addressReceived(address, confs, function(err, _received) {
+    blockchainUtil.addressReceived(address, confs, function(err, _received) {
       if (err) logger.error(err);
       if (_received > 0) received = _received;
       setTimeout(_cb, interval);
